[fix] [#111] 축제 바텀시트의 온보딩 툴팁의 화살표 방향이 위로 향하고 있는 문제
Tooltip designsystem 모듈내에 component 로 이동
관심축제를 선택할 경우 축제 바텀시트 온보딩이 종료되도록 구현
