+++ feature/home/src/main/kotlin/com/unifest/android/feature/home/HomeScreen.kt
@@ -213,6 +213,7 @@ internal fun HomeScreen(
                 isEditMode = uiState.isEditMode,
                 isLikedFestivalDeleteDialogVisible = uiState.isLikedFestivalDeleteDialogVisible,
                 onFestivalUiAction = onFestivalUiAction,
+                isOnboardingCompleted = uiState.isFestivalOnboardingCompleted
             )
         }
     }
