+++ feature/map/src/main/kotlin/com/unifest/android/feature/map/viewmodel/MapUiState.kt
@@ -25,8 +25,8 @@ data class MapUiState(
     val isEditMode: Boolean = false,
     val isPopularMode: Boolean = false,
     val isBoothSelectionMode: Boolean = false,
+    val isFestivalOnboardingCompleted: Boolean = false,
     val isMapOnboardingCompleted: Boolean = false,
-    val isFestivalOnboardingCompleted: Boolean = true,
     val isFestivalSearchBottomSheetVisible: Boolean = false,
     val isLikedFestivalDeleteDialogVisible: Boolean = false,
     val isServerErrorDialogVisible: Boolean = false,
