+++ feature/menu/src/main/kotlin/com/unifest/android/feature/menu/MenuScreen.kt
@@ -307,6 +307,7 @@ fun MenuScreen(
                 isSearchMode = uiState.isSearchMode,
                 isEditMode = uiState.isEditMode,
                 onFestivalUiAction = onFestivalUiAction,
+                isOnboardingCompleted = uiState.isFestivalOnboardingCompleted
             )
         }
     }
